requirements.txt
streamlit>=1.30.0
requests>=2.31.0
anthropic>=0.40.0
pydantic>=2.5.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
nh3>=0.2.14
pandas>=2.0.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
